requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
openpyxl>=3.1.0
numpy>=1.26.0
requests>=2.32.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.1
tqdm>=4.66.0
